remove cli/ops/files.rs; move ops to different mods
