+++ cli/ops/mod.rs
@@ -11,7 +11,6 @@ pub use dispatch_minimal::MinimalOp;
 pub mod compiler;
 pub mod errors;
 pub mod fetch;
-pub mod files;
 pub mod fs;
 pub mod fs_events;
 pub mod io;
